Rename keystore and integrate into app
