+++ src-tauri/src/keystore.rs
